Move api package to submodule
