+++ api/src/main/kotlin/io/github/populus_omnibus/vikbot/api/DefaultMap.kt
