+++ bot/src/main/kotlin/io/github/populus_omnibus/vikbot/Extension.kt
@@ -1,0 +1,9 @@
+package io.github.populus_omnibus.vikbot
+
+import io.github.populus_omnibus.vikbot.api.MaintainEvent
+
+
+operator fun VikBotHandler.plusAssign(event: MaintainEvent<*, *>) {
+    maintainEvent += event.generateTimerEvent()
+    shutdownEvent += event.generateShutdownEvent()
+}
