+++ bot/src/main/kotlin/io/github/populus_omnibus/vikbot/bot/modules/roleselector/RoleSelectorModule.kt
@@ -6,9 +6,9 @@ import io.github.populus_omnibus.vikbot.api.annotations.Module
 import io.github.populus_omnibus.vikbot.api.commands.SlashOptionType
 import io.github.populus_omnibus.vikbot.api.createMemory
 import io.github.populus_omnibus.vikbot.api.maintainEvent
-import io.github.populus_omnibus.vikbot.api.plusAssign
 import io.github.populus_omnibus.vikbot.bot.modules.roleselector.RoleSelectorModule.interactionDeletionWarning
 import io.github.populus_omnibus.vikbot.db.*
+import io.github.populus_omnibus.vikbot.plusAssign
 import net.dv8tion.jda.api.EmbedBuilder
 import net.dv8tion.jda.api.entities.Message
 import net.dv8tion.jda.api.entities.MessageEmbed
@@ -26,7 +26,6 @@ import org.jetbrains.exposed.sql.transactions.TransactionManager
 import org.jetbrains.exposed.sql.transactions.transaction
 import org.slf4j.kotlin.error
 import kotlin.time.Duration.Companion.minutes
-
 
 
 object RoleSelectorModule {
