+++ bot/src/main/kotlin/io/github/populus_omnibus/vikbot/bot/modules/tags/Tags.kt
@@ -9,9 +9,9 @@ import io.github.populus_omnibus.vikbot.api.commands.adminOnly
 import io.github.populus_omnibus.vikbot.api.createMemory
 import io.github.populus_omnibus.vikbot.api.interactions.IdentifiableInteractionHandler
 import io.github.populus_omnibus.vikbot.api.maintainEvent
-import io.github.populus_omnibus.vikbot.api.plusAssign
 import io.github.populus_omnibus.vikbot.bot.isAdmin
 import io.github.populus_omnibus.vikbot.db.*
+import io.github.populus_omnibus.vikbot.plusAssign
 import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.launch
 import kotlinx.datetime.Clock
